exercise 4.21

+++ part4/controllers/blogs.js
@@ -93,14 +93,29 @@ blogsRouter.post("/", async (request, response) => {
 
 blogsRouter.delete("/:id", async (request, response, next) => {
   try {
-    await Blog.findByIdAndDelete(request.params.id);
-    response
-      .status(204)
-      .json({
-        status: "success",
-        data: null
-      })
-      .end();
+    const decodedToken = jwt.verify(request.token, process.env.SECRET);
+    if (!request.token || !decodedToken.id) {
+      return response.status(401).json({ error: "token missing or invalid" });
+    }
+
+    const loggedInUser = await User.findById(decodedToken.id);
+
+    const blog = await Blog.findById(request.params.id);
+
+    if (blog.user.toString() === loggedInUser._id.toString()) {
+      await Blog.findByIdAndDelete(request.params.id);
+      response
+        .status(204)
+        .json({
+          status: "success",
+          data: null
+        })
+        .end();
+    } else {
+      return response
+        .status(401)
+        .json({ error: "you are not the creator of this blog post" });
+    }
   } catch (err) {
     next(err);
   }
